+++ source/Internal/Framework/Module/Command/ModuleActivateCommand.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Command;
 
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ShopConfigurationDaoInterface;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ModuleConfigurationDaoInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Service\ModuleActivationServiceInterface;
 use OxidEsales\EshopCommunity\Internal\Transition\Utility\ContextInterface;
 use Symfony\Component\Console\Command\Command;
@@ -26,7 +26,7 @@ class ModuleActivateCommand extends Command
     public const MESSAGE_MODULE_NOT_FOUND = 'Module - "%s" not found.';
 
     public function __construct(
-        private ShopConfigurationDaoInterface $shopConfigurationDao,
+        private ModuleConfigurationDaoInterface $moduleConfigurationDao,
         private ContextInterface $context,
         private ModuleActivationServiceInterface $moduleActivationService
     ) {
@@ -76,10 +76,6 @@ class ModuleActivateCommand extends Command
      */
     private function isInstalled(string $moduleId): bool
     {
-        $shopConfiguration = $this->shopConfigurationDao->get(
-            $this->context->getCurrentShopId()
-        );
-
-        return $shopConfiguration->hasModuleConfiguration($moduleId);
+        return $this->moduleConfigurationDao->exists($moduleId, $this->context->getCurrentShopId());
     }
 }