+++ source/Internal/Framework/Module/Configuration/Bridge/ModuleConfigurationDaoBridge.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Bridge;
 
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ModuleConfigurationDaoInterface;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ShopEnvironmentConfigurationDaoInterface;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ModuleEnvironmentConfigurationDaoInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration;
 use OxidEsales\EshopCommunity\Internal\Transition\Utility\ContextInterface;
 
@@ -19,7 +19,7 @@ class ModuleConfigurationDaoBridge implements ModuleConfigurationDaoBridgeInterf
     public function __construct(
         private ContextInterface $context,
         private ModuleConfigurationDaoInterface $moduleConfigurationDao,
-        private ShopEnvironmentConfigurationDaoInterface $shopEnvironmentConfigurationDao
+        private ModuleEnvironmentConfigurationDaoInterface $moduleEnvironmentConfigurationDao
     ) {
     }
 
@@ -38,6 +38,6 @@ class ModuleConfigurationDaoBridge implements ModuleConfigurationDaoBridgeInterf
     public function save(ModuleConfiguration $moduleConfiguration)
     {
         $this->moduleConfigurationDao->save($moduleConfiguration, $this->context->getCurrentShopId());
-        $this->shopEnvironmentConfigurationDao->remove($this->context->getCurrentShopId());
+        $this->moduleEnvironmentConfigurationDao->remove($moduleConfiguration->getId(), $this->context->getCurrentShopId());
     }
 }