+++ source/Internal/Framework/Module/Configuration/DataObject/ModuleTemplateExtensionChain.php
@@ -18,18 +18,7 @@ class ModuleTemplateExtensionChain extends RecursiveArrayIterator
     public function getTemplateLoadingPriority(string $templateName): ModuleIdChain
     {
         return new ModuleIdChain(
-            $this->offsetGet($this->normalizeTemplateName($templateName)) ?? []
+            $this->offsetExists($templateName) ? $this->offsetGet($templateName) : []
         );
-    }
-
-    /**
-     * SymfonyConfig replaces hyphens with underscores for ArrayNode keys
-     * @see \Symfony\Component\Config\Definition\ArrayNode::preNormalize
-     */
-    private function normalizeTemplateName(string $templateName): string
-    {
-        return str_contains($templateName, '-') && !str_contains($templateName, '_')
-            ? str_replace('-', '_', $templateName)
-            : $templateName;
     }
 }