+++ tests/Integration/Internal/Framework/Module/Configuration/Dao/ModuleEnvironmentConfigurationDaoTest.php
@@ -10,28 +10,24 @@ declare(strict_types=1);
 namespace OxidEsales\EshopCommunity\Tests\Integration\Internal\Framework\Module\Configuration\Dao;
 
 use OxidEsales\EshopCommunity\Internal\Framework\Storage\FileStorageFactoryInterface;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ShopEnvironmentConfigurationDaoInterface;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ModuleEnvironmentConfigurationDaoInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataMapper\ModuleConfiguration\ModuleSettingsDataMapper;
 use OxidEsales\EshopCommunity\Internal\Transition\Utility\ContextInterface;
 use OxidEsales\EshopCommunity\Tests\ContainerTrait;
 use PHPUnit\Framework\TestCase;
 
-final class ShopEnvironmentConfigurationDaoTest extends TestCase
+final class ModuleEnvironmentConfigurationDaoTest extends TestCase
 {
     use ContainerTrait;
 
     public function testGet(): void
     {
         $this->prepareTestEnvironmentShopConfigurationFile();
-        $environmentConfiguration = $this->get(ShopEnvironmentConfigurationDaoInterface::class)->get(1);
+        $environmentConfiguration = $this->get(ModuleEnvironmentConfigurationDaoInterface::class)->get('testModuleId', 1);
         $expectedEnvironmentConfiguration = [
-            'modules' => [
-                'testModuleId' => [
-                    ModuleSettingsDataMapper::MAPPING_KEY => [
-                        'settingToOverwrite' => [
-                            'value' => 'overwrittenValue',
-                        ]
-                    ]
+            ModuleSettingsDataMapper::MAPPING_KEY => [
+                'settingToOverwrite' => [
+                    'value' => 'overwrittenValue',
                 ]
             ]
         ];
@@ -43,9 +39,9 @@ final class ShopEnvironmentConfigurationDaoTest extends TestCase
     {
         $this->prepareTestEnvironmentShopConfigurationFile();
 
-        $this->get(ShopEnvironmentConfigurationDaoInterface::class)->remove(1);
+        $this->get(ModuleEnvironmentConfigurationDaoInterface::class)->remove('testModuleId', 1);
 
-        $environmentConfiguration = $this->get(ShopEnvironmentConfigurationDaoInterface::class)->get(1);
+        $environmentConfiguration = $this->get(ModuleEnvironmentConfigurationDaoInterface::class)->get('testModuleId', 1);
 
         $this->assertEquals([], $environmentConfiguration);
     }
@@ -56,10 +52,10 @@ final class ShopEnvironmentConfigurationDaoTest extends TestCase
     public function testRemoveOverwriteAlreadyBackupEnvironmentFile(): void
     {
         $this->prepareTestEnvironmentShopConfigurationFile();
-        $this->get(ShopEnvironmentConfigurationDaoInterface::class)->remove(1);
+        $this->get(ModuleEnvironmentConfigurationDaoInterface::class)->remove('testModuleId', 1);
 
         $this->prepareTestEnvironmentShopConfigurationFile();
-        $this->get(ShopEnvironmentConfigurationDaoInterface::class)->remove(1);
+        $this->get(ModuleEnvironmentConfigurationDaoInterface::class)->remove('testModuleId', 1);
     }
 
     /**
@@ -67,7 +63,7 @@ final class ShopEnvironmentConfigurationDaoTest extends TestCase
      */
     public function testRemoveWithNonExistingEnvironmentFile(): void
     {
-        $this->get(ShopEnvironmentConfigurationDaoInterface::class)->remove(1);
+        $this->get(ModuleEnvironmentConfigurationDaoInterface::class)->remove('testModuleId', 1);
     }
 
     private function prepareTestEnvironmentShopConfigurationFile(): void
@@ -75,17 +71,13 @@ final class ShopEnvironmentConfigurationDaoTest extends TestCase
         $fileStorageFactory = $this->get(FileStorageFactoryInterface::class);
         $storage = $fileStorageFactory->create(
             $this->get(ContextInterface::class)
-                ->getProjectConfigurationDirectory() . 'environment/1.yaml'
+                ->getProjectConfigurationDirectory() . 'environment/shops/1/modules/testModuleId.yaml'
         );
 
         $storage->save([
-            'modules' => [
-                'testModuleId' => [
-                    ModuleSettingsDataMapper::MAPPING_KEY => [
-                        'settingToOverwrite' => [
-                            'value' => 'overwrittenValue',
-                        ]
-                    ]
+            ModuleSettingsDataMapper::MAPPING_KEY => [
+                'settingToOverwrite' => [
+                    'value' => 'overwrittenValue',
                 ]
             ]
         ]);